# fix slow sampling when repetition_penalty is set.